requirements.txt
Pillow
pathlib
exif
